# SIGINEX-paquete-completo-3.14.0/siginex-openapi.yaml
openapi: 3.1.0
info:
  title: SIGINEX — API de Diagnóstico del SGI
  version: 1.1.2
  description: >
    Contrato de la API de diagnóstico del Sistema Inteligente de Gestión Integral (SIGINEX),
    consumible por DQnexus. Expone la creación y consulta de diagnósticos,
    resultados por pilar, recomendaciones, rutas de aprendizaje y el feed de vigilancia
    normativa. Los resultados son orientativos y no sustituyen auditoría formal ni concepto legal.
servers:
  - url: https://api.dqnexus.io/siginex/v1
    description: Producción
  - url: https://sandbox.api.dqnexus.io/siginex/v1
    description: Sandbox
security:
  - oauth2ClientCredentials: []
  - apiKey: []
...
paths:
...
  /diagnosticos/{diagnosticoId}/calcular:
    post:
      tags: [Resultados]
      summary: Calcular el diagnóstico
      description: Ejecuta el motor de scoring y recomendaciones sobre las respuestas registradas y cierra el diagnóstico.
      operationId: calcularDiagnostico
      parameters:
        - $ref: '#/components/parameters/DiagnosticoId'
      responses:
        '200':
          description: Diagnóstico calculado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResultadoDiagnostico' }
        '409':
          description: Diagnóstico incompleto (si se exige completitud)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /kb:
    get:
      tags: [Meta]
      summary: Servir el banco de conocimiento vigente
      description: Devuelve el banco de preguntas (500), pesos y modelo de madurez para hidratar el artefacto. Cacheable por ETag.
      operationId: obtenerKb
      security:
        - oauth2ClientCredentials: [kb:read]
        - apiKey: []
      responses:
        '200':
          description: OK
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbBundle' }
  /market-signals:
    get:
      tags: [Vigilancia]
      summary: Feed de vigilancia de mercado
      description: Señales de mercado y competencia (Benchmark_Market_Watch), filtrables por impacto y fecha.
      operationId: listarMarketSignals
      parameters:
        - name: impacto
          in: query
          schema: { type: string, enum: [alto, medio, bajo] }
        - name: desde
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/MarketSignal' }
                  next_cursor: { type: string, nullable: true }
...
components:
  securitySchemes:
    oauth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.dqnexus.io/oauth/token
          scopes:
            diagnosticos:read: Leer diagnósticos y resultados
            diagnosticos:write: Crear y calcular diagnósticos
            normativa:read: Leer el feed de vigilancia normativa
            kb:read: Leer la base de conocimiento
            market:read: Leer el feed de vigilancia de mercado
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    DiagnosticoId:
      name: diagnosticoId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
...
  schemas:
    KbBundle:
      type: object
      properties:
        kb_version: { type: string }
        total_preguntas: { type: integer }
        modelo_madurez: { type: object, additionalProperties: true }
        modulos:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              nombre: { type: string }
              weight: { type: number }
              preguntas: { type: array, items: { type: object, additionalProperties: true } }
    MarketSignal:
      type: object
      properties:
        id: { type: string }
        tipo: { type: string }
        titulo: { type: string }
        fecha: { type: string }
        impacto: { type: string, enum: [alto, medio, bajo] }
        fuente: { type: string }
        resumen: { type: string }
...
    ResultadoDiagnostico:
      type: object
      properties:
        diagnostico_id: { type: string, format: uuid }
        kb_version: { type: string }
        score_sgi: { type: number }
        nivel_sgi: { type: integer, minimum: 1, maximum: 5 }
        nivel_sgi_nombre:
          type: string
          enum: [Inicial, Basico, Gestionado, Optimizado, Excelencia]
        cumplimiento_global_pct: { type: number }
        madurez_global_1a5: { type: number }
        pilares:
          type: array
          items: { $ref: '#/components/schemas/ResultadoPilar' }
        descargo:
          type: string
          default: Resultado orientativo. No sustituye auditoría formal ni concepto legal.
    ResultadoPilar:
      type: object
      properties:
        skill_id:
          type: string
          enum: [sst, calidad, ambiental, cumplimiento, riesgos, gobierno]
        score: { type: number }
        nivel: { type: integer, minimum: 1, maximum: 5 }
        cumplimiento_pct: { type: number, nullable: true }
        madurez_promedio: { type: number, nullable: true }
        hallazgos:
          type: array
          items: { type: string }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: validation_error }
            message: { type: string }
            details:
              type: array
              items: { type: string }
...